Moved StabilizeTransport to FracNetFlow
